requirements.txt
# Requirements lower bounds listed here are our best effort to keep them up to
# date but we do not test them so no guarantee of having them all correct. If
# you find any incorrect lower bounds, let us know or propose a fix.

# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

keystoneauth1>=3.14.0 # Apache-2.0
netaddr>=0.7.18 # BSD
neutron-lib>=2.16.0 # Apache-2.0
openstacksdk>=0.103.0 # Apache-2.0
oslo.config>=8.0.0 # Apache-2.0
oslo.log>=4.3.0 # Apache-2.0
oslo.messaging>=12.4.0 # Apache-2.0
oslo.serialization>=2.28.1 # Apache-2.0
oslo.utils>=4.5.0 # Apache-2.0
ovs>=2.10.0 # Apache-2.0
ovsdbapp>=2.1.0 # Apache-2.0
pbr>=4.0.0 # Apache-2.0
SQLAlchemy>=1.4.23 # MIT
tenacity>=6.0.0 # Apache-2.0
octavia-lib>=2.2.0 # Apache-2.0
